Uses classes to sort deck names and deck objects
There were a lot of duplicate and almost duplicate code in
AnkiDroid. Often the comparator class was put in some random place,
where it could not easily be found to be used again when needed. So I
tried to make this slightly more consistent by creating three
comparators. One for name of deck. One for json objects representing
decks. One for json objects with a name.

I am also correcting a small bug relating to sorting of decks. The
sorting in allSorted was wrong because A::A should appear before A9,
while it sorted it after. (This is because it considered ":" as a
character, and not a separator).

Luckily, this did not cause real trouble since this method is not even
used to sort decks in the deck picker interface.
